Fixing a few more compiler warnings
TODO: We may need to compress this code base significantly, ie - convert ints to uint8's again, but that should be done deliberately and marked clearly where it has been verified as correct

+++ aws_iot_src/shadow/aws_iot_shadow_records.c
@@ -165,7 +165,7 @@ static bool isAckForMyThingName(const char *pTopicName) {
 
 static int AckStatusCallback(MQTTCallbackParams params) {
 	int32_t tokenCount;
-	int32_t i;
+	uint8_t i;
 	void *pJsonHandler = NULL;
 	char temporaryClientToken[MAX_SIZE_CLIENT_ID_WITH_SEQUENCE];
